#1 Custom HooksでApp.tsxのロジック部分を切り出した
