+++ todo-app/.eslintcache
@@ -1,1 +1,1 @@
-[{"/home/vagrant/source/learning-react/todo-app/src/index.tsx":"1","/home/vagrant/source/learning-react/todo-app/src/App.tsx":"2","/home/vagrant/source/learning-react/todo-app/src/components/AddTodo.tsx":"3","/home/vagrant/source/learning-react/todo-app/src/components/TodoElemment.tsx":"4"},{"size":220,"mtime":1610674243121,"results":"5","hashOfConfig":"6"},{"size":1250,"mtime":1610671420533,"results":"7","hashOfConfig":"6"},{"size":512,"mtime":1610679446009,"results":"8","hashOfConfig":"6"},{"size":374,"mtime":1610664896341,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},"1mvneta",{"filePath":"13","messages":"14","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},{"filePath":"15","messages":"16","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},"/home/vagrant/source/learning-react/todo-app/src/index.tsx",[],["19","20","21","22","23"],"/home/vagrant/source/learning-react/todo-app/src/App.tsx",[],"/home/vagrant/source/learning-react/todo-app/src/components/AddTodo.tsx",[],"/home/vagrant/source/learning-react/todo-app/src/components/TodoElemment.tsx",[],{"ruleId":"24","replacedBy":"25"},{"ruleId":"26","replacedBy":"27"},{"ruleId":"28","replacedBy":"29"},{"ruleId":"30","replacedBy":"31"},{"ruleId":"32","replacedBy":"33"},"lines-around-directive",["34"],"global-require",[],"no-buffer-constructor",[],"no-new-require",[],"no-path-concat",[],"padding-line-between-statements"]
+[{"/home/ubuntu/source/learning-react/todo-app/src/index.tsx":"1","/home/ubuntu/source/learning-react/todo-app/src/App.tsx":"2","/home/ubuntu/source/learning-react/todo-app/src/components/AddTodo.tsx":"3","/home/ubuntu/source/learning-react/todo-app/src/components/TodoElemment.tsx":"4","/home/ubuntu/source/learning-react/todo-app/src/hooks/use-todo.tsx":"5"},{"size":220,"mtime":1612753532142,"results":"6","hashOfConfig":"7"},{"size":598,"mtime":1612755129502,"results":"8","hashOfConfig":"7"},{"size":512,"mtime":1612753532132,"results":"9","hashOfConfig":"7"},{"size":374,"mtime":1612753532142,"results":"10","hashOfConfig":"7"},{"size":1257,"mtime":1612755350622,"results":"11","hashOfConfig":"7"},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},"188yblk",{"filePath":"15","messages":"16","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/ubuntu/source/learning-react/todo-app/src/index.tsx",[],["23","24","25","26","27"],"/home/ubuntu/source/learning-react/todo-app/src/App.tsx",[],"/home/ubuntu/source/learning-react/todo-app/src/components/AddTodo.tsx",[],"/home/ubuntu/source/learning-react/todo-app/src/components/TodoElemment.tsx",[],"/home/ubuntu/source/learning-react/todo-app/src/hooks/use-todo.tsx",[],{"ruleId":"28","replacedBy":"29"},{"ruleId":"30","replacedBy":"31"},{"ruleId":"32","replacedBy":"33"},{"ruleId":"34","replacedBy":"35"},{"ruleId":"36","replacedBy":"37"},"lines-around-directive",["38"],"global-require",[],"no-buffer-constructor",[],"no-new-require",[],"no-path-concat",[],"padding-line-between-statements"]
